update server 2

+++ VillageCrawler/Commands/UpdateServerListCommand.cs
@@ -1,5 +1,6 @@
 ﻿using MediatR;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using VillageCrawler.DbContexts;
 using VillageCrawler.Entities;
@@ -7,30 +8,49 @@ using VillageCrawler.Models.Options;
 
 namespace VillageCrawler.Commands
 {
-    public record UpdateServerListCommand(IList<Server> Servers) : IRequest;
+    public record UpdateServerListCommand(List<Server> Servers) : IRequest;
 
-    public class UpdateServerListCommandHandler(IOptions<ConnectionStrings> connections) : IRequestHandler<UpdateServerListCommand>
+    public class UpdateServerListCommandHandler(IOptions<ConnectionStrings> connections, ILogger<UpdateServerListCommand> logger) : IRequestHandler<UpdateServerListCommand>
     {
         private readonly ConnectionStrings _connections = connections.Value;
+        private readonly ILogger<UpdateServerListCommand> _logger = logger;
 
         public async Task Handle(UpdateServerListCommand request, CancellationToken cancellationToken)
         {
             using var context = new ServerDbContext(_connections.Server);
             await context.Database.EnsureCreatedAsync(cancellationToken);
-            await context.BulkMergeAsync(request.Servers);
 
             var servers = await context.Servers
-                .Where(x => x.LastUpdate < DateTime.Now.AddDays(-7))
-                .Select(x => x.Url)
                 .ToListAsync(cancellationToken);
-
-            await context.Servers
-                .Where(x => x.LastUpdate < DateTime.Now.AddDays(-7))
-                .ExecuteDeleteAsync(cancellationToken);
 
             foreach (var server in servers)
             {
-                await DeleteVillageDatabase(server, cancellationToken);
+                var newServer = request.Servers.Find(x => x.Url == server.Url);
+                if (newServer is null) continue;
+
+                server.AllianceCount = newServer.AllianceCount;
+                server.PlayerCount = newServer.PlayerCount;
+                server.VillageCount = newServer.VillageCount;
+                server.LastUpdate = newServer.LastUpdate;
+            }
+
+            await context.BulkSaveChangesAsync(cancellationToken);
+
+            var timeoutServers = await context.Servers
+                .Where(x => x.LastUpdate < DateTime.Now.AddDays(-7))
+                .Select(x => new { x.Id, x.Url })
+                .ToListAsync(cancellationToken);
+
+            if (timeoutServers.Count == 0) return;
+            _logger.LogInformation("Deleting {Count} servers: {Servers}", timeoutServers.Count, timeoutServers);
+
+            await context.Servers
+                .WhereBulkContains(timeoutServers.Select(x => x.Id))
+                .ExecuteDeleteAsync(cancellationToken);
+
+            foreach (var server in timeoutServers)
+            {
+                await DeleteVillageDatabase(server.Url, cancellationToken);
             }
         }
